==> src/3a-ML_Algos_Multilabel_to_Multiclass.ipynb <==
# Part 3a: Machine Learning - Multi-label Converted to Multi-class Problem

import pandas as pd
import numpy as np
import pickle
from sklearn.pipeline import  Pipeline
from sklearn.model_selection import train_test_split, KFold, GridSearchCV, ParameterGrid
from sklearn.preprocessing import StandardScaler#,  MinMaxScaler
from sklearn.metrics import f1_score,  precision_score, recall_score

from skmultilearn.problem_transform import LabelPowerset

from sklearn.dummy import DummyClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.multiclass import OneVsRestClassifier

features = pd.read_csv('../data/features_plus_descriptions.csv')
features.set_index('Feature Type and Number', inplace=True)
features.drop(['S5', 'D21'], axis=0, inplace=True)
print('number of features:', len(features.index))

target = ['H0_Best','H1_Best', 'H2_Best', 'H3_Best', 'H4_Best', 'H5_Best']

df = pd.read_csv('../data/multilabel_with_tol_raw_data.csv')
X = df.loc[:, features.index]
y = df[target]
y[:14] # Display the target columns

#### Use standard scaler for all features

def run_pipeline(X, y, param_grid):
    test_scores = []
    best_models = []
    for i in range(10):
        random_state = 431 * i
        X_other, X_test, y_other, y_test = train_test_split(X, y, test_size = 0.2, random_state=random_state)
        kf = KFold(n_splits=4, shuffle=True, random_state=random_state)
        pipe = Pipeline([
            ("preprocessor", StandardScaler()), 
            ("clf", LabelPowerset())
        ])
        grid = GridSearchCV(pipe, param_grid, scoring="f1_micro", cv=kf, return_train_score=True, verbose=5, n_jobs=-1)   
        grid.fit(X_other, y_other)
        y_pred = grid.predict(X_test)
        prec = precision_score(y_test, y_pred, average="micro")
        rec = recall_score(y_test, y_pred, average="micro")
        f1 = f1_score(y_test, y_pred, average="micro")
        print("Best Params: {} \nBest CV Score: {}".format(grid.best_params_, np.round(grid.best_score_, decimals=6)))
        print("Precision:     {}\nRecall:        {}\nf1_micro:      {} \n".format(np.round(prec, decimals=6), np.round(rec, decimals=6),np.round(f1, decimals=6)))
        test_scores.append(f1)
        best_models.append(grid)
    return test_scores, best_models

## Dummy Classifier (Baseline Test Score)

param_grid = { 'clf': [DummyClassifier()],
               'clf__strategy': ["stratified", "most_frequent", "prior", "uniform"]}
test_scores_dummy, best_models_dummy = run_pipeline(X, y, param_grid)
amax = np.argmax(test_scores_dummy)
print('best test score for DummyClassifier:', test_scores_dummy[amax])
print('best params for DummyClassifier:', best_models_dummy[amax].best_params_)

print('mean test score for Dummy Classifier:', np.mean(test_scores_dummy) )
print('stdev of test score for Dummy Classifier:', np.std(test_scores_dummy))

## Random Forest Classifier

param_grid = {  'clf': [RandomForestClassifier()],
                'clf__max_depth': [15, 20, 25, 30, 35, 40],
                'clf__max_features': [15, 20, 25, 30, 35, 40, 45, 51]} 

test_scores_RF, best_models_RF = run_pipeline(X, y, param_grid)


amax = np.argmax(test_scores_RF)
print(amax)
print('best test score for Random Forest:', test_scores_RF[amax])
print('best estimator for Random Forest:', best_models_RF[amax].best_params_)

print('mean test score for Random Forest:', np.mean(test_scores_RF) )
print('stdev of test score for Random Forest:', np.std(test_scores_RF))

output = open('../results/RF_models_micro.pkl', 'wb')
pickle.dump(best_models_RF, output)
output.close()

## K Nearest Neighbors Classifier

param_grid = {'clf': [KNeighborsClassifier()], 
              'clf__n_neighbors': [5,10,15,20,30,40,50,60],
              'clf__weights': ['distance', 'uniform'] }

test_scores_KNN, best_models_KNN = run_pipeline(X, y, param_grid)
print(np.mean(test_scores_KNN), '+/-', np.std(test_scores_KNN))

amax = np.argmax(test_scores_KNN)
print(amax)
print('best test score for K Nearest Neighbors:', test_scores_KNN[amax])
print('best params for K Nearest Neighbors:', best_models_KNN[amax].best_params_)

print('mean test score for K Nearest Neighbors:', np.mean(test_scores_KNN) )
print('stdev of test score for K Nearest Neighbors:', np.std(test_scores_KNN))

output = open('../results/KNN_models_micro.pkl', 'wb')
pickle.dump(best_models_KNN, output)
output.close()

## SVC

param_grid = { 'clf': [OneVsRestClassifier(SVC(max_iter=100000000, cache_size=3000))],
               'clf__estimator__C': np.logspace(1,4,num=4),
               'clf__estimator__tol': np.logspace(-4, -2,num=3) } 

test_scores_SVC, best_models_SVC = run_pipeline(X, y, param_grid)

amax = np.argmax(test_scores_SVC)
print(amax)
print('best test score for SVC OneVsRest:', test_scores_SVC[amax])
print('best estimator for SVC OneVsRest:', best_models_SVC[amax].best_params_)

print('mean test score for SVC OneVsRest:', np.mean(test_scores_SVC) )
print('stdev of test score for SVC OneVsRest:', np.std(test_scores_SVC))

output = open('../results/SVC_models_micro.pkl', 'wb')
pickle.dump(best_models_SVC, output)
output.close()

## Logistic Regression

param_grid = { 'clf': [OneVsRestClassifier(LogisticRegression(max_iter=100000, 
                                                              warm_start=True, 
                                                              multi_class='ovr'))],
               'clf__estimator__penalty': ['l2'],
               'clf__estimator__C': np.logspace(0, 4, 5),
               'clf__estimator__solver': ['sag', 'lbfgs'] }

test_scores_LR, best_models_LR = run_pipeline(X, y, param_grid)

amax = np.argmax(test_scores_LR)
print(amax)
print('best test score for Log Reg:', test_scores_LR[amax])
print('best params for Log Reg:', best_models_LR[amax].best_params_)

print('mean test score for Log Reg:', np.mean(test_scores_LR) )
print('stdev of test score for Log Reg:', np.std(test_scores_LR))

output = open('../results/LR_models_micro.pkl', 'wb')
pickle.dump(best_models_LR, output)
output.close()
