==> src/Part_4-Merge_Target_Classes.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pylab as plt
from itertools import chain, combinations, permutations
from skmultilearn.problem_transform import LabelPowerset
from sklearn.dummy import DummyClassifier

from collections import OrderedDict

features = pd.read_csv('../data/features_plus_descriptions.csv')
features.set_index('Feature Type and Number', inplace=True)

df_multilabel_strict = pd.read_csv('../data/multilabel_raw_data_strict.csv')
df_multiclass = pd.read_csv('../data/multiclass_target_raw_data.csv')
H_Best = ['H0_Best', 'H1_Best', 'H2_Best', 'H3_Best', 'H4_Best', 'H5_Best']

y_strict = df_multilabel_strict[H_Best] 
y_best = df_multiclass['Best Heuristic']


X_strict = df_multilabel_strict.drop(H_Best, axis=1)

X_orig = df_multiclass.drop(H_Best + ['Best Heuristic'], axis=1)
X_orig.head()

y_strict

np.array(y_strict)

def get_unique_combos(X, y):
    lp = LabelPowerset(DummyClassifier(strategy='stratified'))
    lp.fit(X, y)
    return lp.unique_combinations_

def print_mapping(dictionary):
    for key in dictionary.keys():
        print('{} : {}'.format(key, dictionary[key]))
        
def str_keys_to_list(string):
    return [int(num) for num in string.split(',')]

unique_strict = get_unique_combos(X_strict, y_strict)

def get_powerset_counts(target, powerset):
    store_counts = {combo: 0 for combo in powerset.keys()}
    for combo in powerset.keys():
        int_list = str_keys_to_list(combo)
        pattern = np.zeros(6,dtype=int)
        for i in int_list:
            pattern[i] = 1
        for a in range(len(target)):
            if np.array_equal(target[a], pattern):
                store_counts[combo] += 1
    return store_counts


arr_y_strict = np.array(y_strict)
counts_strict = get_powerset_counts(arr_y_strict, unique_strict)

sorted_counts_strict = OrderedDict(sorted(counts_strict.items(), key= lambda k: k[1]))
counts_strict_df = pd.DataFrame(sorted_counts_strict.values(), index=sorted_counts_strict.keys(), columns=['strict'])
print(counts_strict_df)

print_mapping(unique_strict)

np.array(y_strict)
print(np.array(y_strict))

def modify_target_labels(target, powerset):
    N = len(target)
    target_newlabels = np.zeros(shape=(N,8), dtype=int)
    count = 0
    for combo in powerset.keys():
        int_list = str_keys_to_list(combo)
        int_list_copy = str_keys_to_list(combo)
        if 1 in int_list and 2 in int_list:
            int_list_copy.remove(1)
            int_list_copy.remove(2)
            int_list_copy.append(6)
        if 4 in int_list and 5 in int_list:
            int_list_copy.remove(4)
            int_list_copy.remove(5)
            int_list_copy.append(7)
        pattern = np.zeros(6,dtype=int)
        for i in int_list:
            pattern[i] = 1
        for a in range(N):
            if np.array_equal(target[a], pattern):
                for j in int_list_copy:
                    target_newlabels[a][j] = 1
                count += 1
                
    print(count)
    return target_newlabels


target_newlabels = modify_target_labels(np.array(y_strict), unique_strict)

target_newlabels = modify_target_labels(y_strict.to_numpy(), unique_strict)

target_newlabels

target_new = pd.DataFrame(target_newlabels)

target_new

def get_unique_combos(X, y):
    lp = LabelPowerset(DummyClassifier(strategy='stratified'))
    lp.fit(X, y)
    return lp.unique_combinations_

unique_new = get_unique_combos(X_strict, target_new)

print_mapping(unique_new)

def get_powerset_counts_new(target, powerset):
    store_counts = {combo: 0 for combo in powerset.keys()}
    for combo in powerset.keys():
        int_list = str_keys_to_list(combo)
        pattern = np.zeros(8,dtype=int)
        for i in int_list:
            pattern[i] = 1
        for a in range(len(target)):
            if np.array_equal(target[a], pattern):
                store_counts[combo] += 1
    return store_counts

counts_new = get_powerset_counts_new(np.array(target_new), unique_new)
sorted_counts_new = OrderedDict(sorted(counts_new.items(), key= lambda k: k[1]))

counts_new_df = pd.DataFrame(sorted_counts_new.values(),index=sorted_counts_new.keys(), columns=['strict_new'])
counts_new_df

def modify_target_labels_round2(target, powerset):
    N = len(target)
    target_newlabels = np.zeros(shape=(N,9), dtype=int)
    count = 0
    for combo in powerset.keys():
        int_list = str_keys_to_list(combo)
        int_list_copy = str_keys_to_list(combo)
        if 3 in int_list and 7 in int_list:
            int_list_copy.remove(3)
            int_list_copy.remove(7)
            int_list_copy.append(8)
        pattern = np.zeros(8,dtype=int)
        for i in int_list:
            pattern[i] = 1
        for a in range(N):
            if np.array_equal(target[a], pattern):
                for j in int_list_copy:
                    target_newlabels[a][j] = 1
                count += 1
                
    print(count)
    return target_newlabels


target_new2 = modify_target_labels_round2(target_newlabels, unique_new)
target_new2_df = pd.DataFrame(target_new2)

unique_new2 = get_unique_combos(X_strict, target_new2_df)

def get_powerset_counts_new(target, powerset):
    store_counts = {combo: 0 for combo in powerset.keys()}
    for combo in powerset.keys():
        int_list = str_keys_to_list(combo)
        pattern = np.zeros(9,dtype=int)
        for i in int_list:
            pattern[i] = 1
        for a in range(len(target)):
            if np.array_equal(target[a], pattern):
                store_counts[combo] += 1
    return store_counts

counts_new2 = get_powerset_counts_new(target_new2, unique_new2)
sorted_counts_new2 = OrderedDict(sorted(counts_new2.items(), key= lambda k: k[1]))



sorted_counts_new2

counts_new_df2 = pd.DataFrame(sorted_counts_new2.values(),index=sorted_counts_new2.keys(), columns=['strict_new'])
counts_new_df2

df_merged = X_strict.merge(target_new2_df, left_index=True, right_index=True)

df_merged.head()

df_merged.to_csv('../data/raw_data_modified_classes.csv', index=None)
